Add user, project route

+++ src/handlers/firebaseUpload.ts
@@ -1,11 +1,13 @@
 import multer from 'multer'
-import { v4 as uuidv4 } from 'uuid'
 import admin from '@/configs/firebase-admin'
 
 const upload = multer({ storage: multer.memoryStorage() })
 const bucket = admin.storage().bucket()
 
-const singleUpload = async (file: any, destination: string) => {
+export const singleUpload = async (
+  file: any,
+  destination: string
+): Promise<null | { name: string; minetype: string; url: string }> => {
   return new Promise((resolve, reject) => {
     if (!file) {
       return resolve(null)
@@ -41,7 +43,7 @@ const singleUpload = async (file: any, destination: string) => {
   })
 }
 
-const multipleUpload = async (files: any, destination: string) => {
+export const multipleUpload = async (files: any, destination: string) => {
   const promises = files.map(async (file: any) => {
     return await singleUpload(file, destination)
   })
@@ -51,7 +53,7 @@ const multipleUpload = async (files: any, destination: string) => {
   return urls
 }
 
-const deleteFileStorageByFileName = async (destination: string, fileName: string) => {
+export const deleteFileStorageByFileName = async (destination: string, fileName: string) => {
   try {
     await bucket.file(`${destination}/${fileName}`).delete()
   } catch (error: any) {
@@ -59,16 +61,18 @@ const deleteFileStorageByFileName = async (destination: string, fileName: string
   }
 }
 
-const deleteFileStorageByUrl = async (url: string) => {
+export const deleteFileStorageByUrl = async (url: string) => {
   try {
-    const filePath = url.split(`https://storage.googleapis.com/${bucket.name}/`)[1]
-    await bucket.file(filePath).delete()
+    if (url) {
+      const filePath = url.split(`https://storage.googleapis.com/${bucket.name}/`)[1]
+      await bucket.file(filePath).delete()
+    }
   } catch (error: any) {
     console.log(error.message)
   }
 }
 
-const deleteFolderStorage = async (destination: string) => {
+export const deleteFolderStorage = async (destination: string) => {
   try {
     await bucket.deleteFiles({ prefix: destination })
   } catch (error: any) {
@@ -76,7 +80,7 @@ const deleteFolderStorage = async (destination: string) => {
   }
 }
 
-const getFileUrls = async (destination: string) => {
+export const getFileUrls = async (destination: string) => {
   try {
     const [files] = await bucket.getFiles({
       prefix: destination
@@ -102,12 +106,4 @@ const getFileUrls = async (destination: string) => {
   }
 }
 
-export default {
-  upload,
-  singleUpload,
-  multipleUpload,
-  deleteFileStorageByFileName,
-  deleteFileStorageByUrl,
-  deleteFolderStorage,
-  getFileUrls
-}
+export default upload
